requirements.txt
langchain
openai
pandas
google-generativeai
